auto 0 19 243

+++ src/smallstring.h
@@ -1,4 +1,6 @@
 #include "mystring.h"
 
-void smallstring_write(int out, const char* s, size_t len);
+#include <stdint.h>
+
+void smallstring_write(int out, const char* s, uint16_t len);
 void smallstring_read(string* out, int inp);
